+++ client/src/components/home/Card.js
@@ -1,42 +1,26 @@
 import React, { useState, useEffect } from 'react';
-import axios from 'axios';
 import styled from 'styled-components';
 import Modal from '../modal/Modal';
 import useModal from '../../hooks/useModal';
 import PictureShow from '../picture/PictureShow';
-import defaultUserImage from '../../img/defaultUserImage.png'
 import commentsImage from '../../img/comments.png'
 import viewsImage from '../../img/views.png'
 import { ImageConsumer } from '../../providers/ImageProvider'
 
-// const Card = ( { category_id, title, id, url, user_id, views, description, fetchUser, user, userImage }, ) => {
 const Card = (props) => {
-
   const { open, toggle } = useModal();
-  // const [commentsLength, setComments] = useState(0)
-  const [firstPic, setFirstPic] = useState(true)
-  const [nextPic, setNextPic ] = useState([])
   const id = props.image.id
-  // const user = props.user
   const user_id = props.image.user_id
-  const category_id = props.image.category_id
-  const title = props.image.title
+  const [views, setViews] = useState(props.image.views)
   const url = props.image.url
-  const description = props.image.description
-  const views = props.image.views
-  // const userImage = props.userImage
-  // const commentsLength = props.commentsLength
 
   const [user, setUser] = useState("");
-  const [userImage, setUserImage] = useState();
   const [comments, setComments] = useState([]);
   
-
   useEffect(() => {
     props.fetchUser(user_id)
       .then (res => {
         setUser(res.data)
-        setUserImage(res.data.image)
       })
       .catch(console.log)
 
@@ -48,15 +32,7 @@ const Card = (props) => {
 
   }, [])
 
-
-  // const switchPicture = (incomingPic, incomingComments) => {
-  //   setFirstPic(false)
-  //   setNextPic(incomingPic)
-    // props.fetchUser(incomingPic.user_id)
-        
-      
-  
- //Change back to first pic on close
+  const updateViewsState = () => setViews(views + 1)
 
   const toggleAndSetId = () => {
     props.setImageId(id)
@@ -65,27 +41,14 @@ const Card = (props) => {
   return (
     <CardBorder>
       <Modal onClose={toggle} open={open}>     
-        
-          <PictureShow 
-            toggle={toggle}
-            user={user}
-            user_id={user_id} 
-            category_id={category_id} 
-            title={title} 
-            // id={id} 
-            url={url} 
-            description={description} 
-            views={views}
-            // comments={comments}
-          />    
-        
+          <PictureShow updateViews={updateViewsState} />   
       </Modal>       
       <CardDiv>
         <StyledImage src={url} onClick={toggleAndSetId} />
       </CardDiv>
       <PointerOff>
         <CardFooterLeft>
-          <SmallImage image={userImage}/>
+          <SmallImage image={user.image}/>
           {user.first_name} 
         </CardFooterLeft>
         <CardFooterRight>
